Added New Stuff
More logic and menu items, the total text view also dynamically updates as item counts are increased. The string detailing the items, cost, and total is saved for use with the intent that will be implemented later. Also need to add Drinks menu items as well as the submit button at the end.